backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis 和 Celery
redis==5.0.1
celery==5.3.4

# PDF 处理
PyMuPDF==1.23.8
pdf2image==3.1.0
Pillow==10.1.0
opencv-python==4.8.1.78

# AI 模型调用
httpx==0.25.2
aiohttp==3.9.1

# 数据验证和配置
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
